# chore: adjust default jmh settings for benchmarks (MINOR)